+++ monad-eth-txpool/src/lib.rs
@@ -25,7 +25,6 @@ use crate::{
 };
 
 mod pool;
-pub mod test_utils;
 mod transaction;
 mod utils;
