+++ monad-rpc/src/websocket.rs
@@ -118,7 +118,7 @@ mod tests {
             max_response_size: 25_000_000,
             allow_unprotected_txs: false,
             rate_limiter: Arc::new(Semaphore::new(1000)),
-            tx_pool: Arc::new(tokio::sync::Mutex::new(vpool::VirtualPool::new(
+            tx_pool: Arc::new(tokio::sync::Mutex::new(monad_eth_vpool::VirtualPool::new(
                 ipc_sender, 20_000, None, None,
             ))),
         };
